# contracts/alchemy-openapi.json
{
  "components": {
    "schemas": {
...
      "DataStatus": {
        "enum": [
          "verified",
          "source_reported",
          "conflicted",
          "demo",
          "incomplete"
        ],
        "title": "DataStatus",
        "type": "string"
      },
...
      "DisabledInquiryRequest": {
        "additionalProperties": false,
        "properties": {
          "query": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "DisabledInquiryRequest",
        "type": "object"
      },
      "DistributionEntry": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "proportion": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Proportion",
            "type": "number"
          },
          "value": {
            "title": "Value",
            "type": "string"
          },
          "weightGrams": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weightgrams"
          },
          "weightedProportion": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weightedproportion"
          }
        },
        "required": [
          "value",
          "count",
          "proportion"
        ],
        "title": "DistributionEntry",
        "type": "object"
      },
...
      "Envelope_FormulaAnalysisResult_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FormulaAnalysisResult"
          },
          "meta": {
            "$ref": "#/components/schemas/KnowledgeMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[FormulaAnalysisResult]",
        "type": "object"
      },
      "Envelope_FormulaComparisonResult_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FormulaComparisonResult"
          },
          "meta": {
            "$ref": "#/components/schemas/KnowledgeMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[FormulaComparisonResult]",
        "type": "object"
      },
...
      "Envelope_RetrievalPackage_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RetrievalPackage"
          },
          "meta": {
            "$ref": "#/components/schemas/KnowledgeMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[RetrievalPackage]",
        "type": "object"
      },
...
      "FormulaAnalysisRequest": {
        "additionalProperties": false,
        "properties": {
          "composition": {
            "$ref": "#/components/schemas/FormulaCompositionInput-Input"
          }
        },
        "required": [
          "composition"
        ],
        "title": "FormulaAnalysisRequest",
        "type": "object"
      },
      "FormulaAnalysisResult": {
        "additionalProperties": false,
        "properties": {
          "algorithmVersion": {
            "title": "Algorithmversion",
            "type": "string"
          },
          "conflicts": {
            "items": {
              "type": "string"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "dataVersion": {
            "title": "Dataversion",
            "type": "string"
          },
          "distributions": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/DistributionEntry"
              },
              "type": "array"
            },
            "title": "Distributions",
            "type": "object"
          },
          "documentedActions": {
            "items": {
              "type": "string"
            },
            "title": "Documentedactions",
            "type": "array"
          },
          "documentedPatterns": {
            "items": {
              "type": "string"
            },
            "title": "Documentedpatterns",
            "type": "array"
          },
          "exactDuplicateIngredientIds": {
            "items": {
              "type": "string"
            },
            "title": "Exactduplicateingredientids",
            "type": "array"
          },
          "explicitRoles": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Explicitroles",
            "type": "object"
          },
          "missingData": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Missingdata",
            "type": "object"
          },
          "normalizedIngredients": {
            "items": {
              "$ref": "#/components/schemas/NormalizedIngredient"
            },
            "title": "Normalizedingredients",
            "type": "array"
          },
          "originalInput": {
            "$ref": "#/components/schemas/FormulaCompositionInput-Output"
          },
          "pairSignals": {
            "items": {
              "$ref": "#/components/schemas/IngredientPairSignal"
            },
            "title": "Pairsignals",
            "type": "array"
          },
          "preparationDistinctions": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Preparationdistinctions",
            "type": "object"
          },
          "reviewStatusBreakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Reviewstatusbreakdown",
            "type": "object"
          },
          "sourceCoveragePercentage": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Sourcecoveragepercentage",
            "type": "number"
          },
          "supportedUnitTotalsGrams": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supportedunittotalsgrams"
          },
          "unsupportedOrUnresolvedUnits": {
            "items": {
              "type": "string"
            },
            "title": "Unsupportedorunresolvedunits",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "algorithmVersion",
          "dataVersion",
          "originalInput",
          "normalizedIngredients",
          "exactDuplicateIngredientIds",
          "preparationDistinctions",
          "supportedUnitTotalsGrams",
          "unsupportedOrUnresolvedUnits",
          "distributions",
          "documentedActions",
          "documentedPatterns",
          "explicitRoles",
          "pairSignals",
          "conflicts",
          "missingData",
          "sourceCoveragePercentage",
          "reviewStatusBreakdown",
          "warnings"
        ],
        "title": "FormulaAnalysisResult",
        "type": "object"
      },
      "FormulaComparisonRequest": {
        "additionalProperties": false,
        "properties": {
          "compositions": {
            "items": {
              "$ref": "#/components/schemas/FormulaCompositionInput-Input"
            },
            "maxItems": 4,
            "minItems": 2,
            "title": "Compositions",
            "type": "array"
          }
        },
        "required": [
          "compositions"
        ],
        "title": "FormulaComparisonRequest",
        "type": "object"
      },
      "FormulaComparisonResult": {
        "additionalProperties": false,
        "properties": {
          "algorithmVersion": {
            "title": "Algorithmversion",
            "type": "string"
          },
          "analyses": {
            "items": {
              "$ref": "#/components/schemas/FormulaAnalysisResult"
            },
            "title": "Analyses",
            "type": "array"
          },
          "combinedDistributions": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/DistributionEntry"
              },
              "type": "array"
            },
            "title": "Combineddistributions",
            "type": "object"
          },
          "completenessSummary": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Completenesssummary",
            "type": "object"
          },
          "crossFormulaPairSignals": {
            "items": {
              "$ref": "#/components/schemas/IngredientPairSignal"
            },
            "title": "Crossformulapairsignals",
            "type": "array"
          },
          "dataVersion": {
            "title": "Dataversion",
            "type": "string"
          },
          "distinctActions": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Distinctactions",
            "type": "object"
          },
          "distinctPatterns": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Distinctpatterns",
            "type": "object"
          },
          "pairwise": {
            "items": {
              "$ref": "#/components/schemas/PairwiseComparison"
            },
            "title": "Pairwise",
            "type": "array"
          },
          "repeatedIngredientIds": {
            "items": {
              "type": "string"
            },
            "title": "Repeatedingredientids",
            "type": "array"
          },
          "sharedActions": {
            "items": {
              "type": "string"
            },
            "title": "Sharedactions",
            "type": "array"
          },
          "sharedPatterns": {
            "items": {
              "type": "string"
            },
            "title": "Sharedpatterns",
            "type": "array"
          },
          "sourceConflicts": {
            "items": {
              "type": "string"
            },
            "title": "Sourceconflicts",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "algorithmVersion",
          "dataVersion",
          "analyses",
          "pairwise",
          "repeatedIngredientIds",
          "combinedDistributions",
          "sharedActions",
          "distinctActions",
          "sharedPatterns",
          "distinctPatterns",
          "crossFormulaPairSignals",
          "sourceConflicts",
          "warnings",
          "completenessSummary"
        ],
        "title": "FormulaComparisonResult",
        "type": "object"
      },
      "FormulaCompositionInput-Input": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/FormulaIngredientInput-Input"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Ingredients",
            "type": "array"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "sourceFormulaId": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceformulaid"
          }
        },
        "required": [
          "name",
          "ingredients"
        ],
        "title": "FormulaCompositionInput",
        "type": "object"
      },
      "FormulaCompositionInput-Output": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/FormulaIngredientInput-Output"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Ingredients",
            "type": "array"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "sourceFormulaId": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceformulaid"
          }
        },
        "required": [
          "name",
          "ingredients"
        ],
        "title": "FormulaCompositionInput",
        "type": "object"
      },
      "FormulaIngredientInput-Input": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "herbMaterialId": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Herbmaterialid",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "preparationId": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preparationid"
          },
          "role": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "sourceText": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcetext"
          },
          "unit": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 1,
                "pattern": "^[\\w%./-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "herbMaterialId"
        ],
        "title": "FormulaIngredientInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeMeta": {
        "additionalProperties": false,
        "properties": {
          "algorithmVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Algorithmversion"
          },
          "dataStatus": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceSummary"
            },
            "title": "Sources",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "requestId",
          "dataStatus",
          "schemaVersion"
        ],
        "title": "KnowledgeMeta",
        "type": "object"
      },
...
      "RetrievalContextRequest": {
        "additionalProperties": false,
        "properties": {
          "entityIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Entityids",
            "type": "array"
          },
          "maximumCharacterBudget": {
            "default": 12000,
            "maximum": 50000.0,
            "minimum": 500.0,
            "title": "Maximumcharacterbudget",
            "type": "integer"
          },
          "maximumPassages": {
            "default": 10,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Maximumpassages",
            "type": "integer"
          },
          "passageIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Passageids",
            "type": "array"
          },
          "query": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "sourceIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Sourceids",
            "type": "array"
          }
        },
        "title": "RetrievalContextRequest",
        "type": "object"
      },
...
      "RetrievalPackage": {
        "additionalProperties": false,
        "properties": {
          "characterBudget": {
            "minimum": 1.0,
            "title": "Characterbudget",
            "type": "integer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "graphNeighborhoodFacts": {
            "items": {
              "$ref": "#/components/schemas/RetrievalFact"
            },
            "title": "Graphneighborhoodfacts",
            "type": "array"
          },
          "matchedEntities": {
            "items": {
              "$ref": "#/components/schemas/EntitySummary"
            },
            "title": "Matchedentities",
            "type": "array"
          },
          "passages": {
            "items": {
              "$ref": "#/components/schemas/PassageRecord"
            },
            "title": "Passages",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "reviewStatuses": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Reviewstatuses",
            "type": "object"
          },
          "sourceStatuses": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Sourcestatuses",
            "type": "object"
          },
          "unresolvedAmbiguities": {
            "items": {
              "type": "string"
            },
            "title": "Unresolvedambiguities",
            "type": "array"
          },
          "usedCharacters": {
            "minimum": 0.0,
            "title": "Usedcharacters",
            "type": "integer"
          }
        },
        "required": [
          "query",
          "passages",
          "citations",
          "matchedEntities",
          "graphNeighborhoodFacts",
          "unresolvedAmbiguities",
          "sourceStatuses",
          "reviewStatuses",
          "usedCharacters",
          "characterBudget"
        ],
        "title": "RetrievalPackage",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Educational and research information only; not medical advice. Data may be incomplete, and absence of a known interaction does not establish safety.",
    "summary": "Sourced graph research and deterministic formula analysis",
    "title": "Current Alchemy API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/formulas/analyze": {
      "post": {
        "operationId": "analyze_api_v1_formulas_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormulaAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_FormulaAnalysisResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "formula analysis"
        ]
      }
    },
    "/api/v1/formulas/compare": {
      "post": {
        "operationId": "compare_api_v1_formulas_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormulaComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_FormulaComparisonResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare",
        "tags": [
          "formula analysis"
        ]
      }
    },
...
    "/api/v1/inquiry/synthesize": {
      "post": {
        "operationId": "disabled_synthesis_api_v1_inquiry_synthesize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisabledInquiryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Disabled Synthesis",
        "tags": [
          "retrieval"
        ]
      }
    },
...
    "/api/v1/retrieval/context": {
      "post": {
        "operationId": "retrieval_context_api_v1_retrieval_context_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrievalContextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RetrievalPackage_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retrieval Context",
        "tags": [
          "retrieval"
        ]
      }
    },
...
  }
}